Big Threejs Update
Made a massive transition over to threejs. Canvas and Threejs have been
further modularized and each one is able to be loaded independently.
ThreejsSM now works with the state manager and allows for different
scene rendering per state. PauseState has a screen with a textured
rotating cube. PlayState has a skybox with an untextured rotating cube
to demonstrate this.